Start refactoring core logic to package
